refactor(midnight): group wallet sync backends under wallet_sync/
Move the three Midnight balance sync backends — shielded (Zswap),
unshielded, and dust — out of the ows-midnight crate root into a single
wallet_sync/ module:

  wallet_sync/
    shielded/{mod,sync_cache,zswap_ledger}.rs
    unshielded.rs
    dust/{mod,sync_cache}.rs

Pure relocation. Module paths change: super:: -> crate:: for the shared
crate-root modules (cache_io, indexer_ws, midnight_env, tip_verify,
ShieldedBalances), and the per-backend cache references repoint to their
new sibling module names. No behavior change; the public API re-exported
from lib.rs is unchanged.
