+++ ows/crates/ows-midnight/src/wallet_sync/mod.rs
@@ -1,0 +1,5 @@
+//! Wallet balance/asset sync backends: shielded (Zswap), unshielded, and dust.
+
+pub mod dust;
+pub mod shielded;
+pub mod unshielded;
